Adding script to calculate the BN priors
This includes modifications to the initial values, as well
as getting rid of the "prior" means and std. There were also
issues with `NaN` popping up from the noise calculation; I've
replaced it with Brett's version from `MolSim`, modified to
calculate the noise mean and std. As far as I can tell, this
no longer throws `NaN`s.

Also porting functions from Ryan's version of `simulate_lte`: that
way we don't have to append paths or anything like that.