Appears to be working codebase!
BN prior generation is validated and seems fine. We had a lot of
trouble with overly confident predictions, which turns out was because
the prior means and std were not being calculated correctly from the
loaded sample chain.

Based on the LOG output, the std in particular is now correctly wide
and should work for 1- and 2-CNN.