Ran 2-CNN locally, and the results look pretty!
Modified the analysis of the chain with semi automated
walker rejection: in the notebook, chains with means 1.2 sigma
away from the ensemble average are excluded from analysis.

With this, the `make_results_2CNN.py` script has also been modified
to load in the analyzed chain and format the statistics output.